Move test package inside the git package
It is not being used by other packages so it should belong to `git` package.
